move division
