+++ src/lib.rs
@@ -30,7 +30,7 @@ pub mod vlq;
 pub use crate::channel::Channel;
 use crate::constants::{FILE_META_EVENT, FILE_SYSEX_F0, FILE_SYSEX_F7};
 use crate::error::LibResult;
-pub use crate::file::{Format, Header};
+pub use crate::file::{Division, Format, Header};
 pub use crate::message::{Message, NoteMessage, NoteNumber, Program, ProgramChangeValue, Velocity};
 use crate::vlq::Vlq;
 pub use error::{Error, Result};
@@ -164,97 +164,6 @@ fn ensure_end_of_track(mut track: Track) -> LibResult<Track> {
         track.push_event(0, Event::Meta(MetaEvent::EndOfTrack))?;
     }
     Ok(track)
-}
-
-/// <division>, specifies the meaning of the delta-times. It has two formats, one for metrical time,
-/// and one for time-code-based time:
-#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd, Hash)]
-pub enum Division {
-    /// If bit 15 of <division> is a zero, the bits 14 thru 0 represent the number of delta-time
-    /// "ticks" which make up a quarter-note. For instance, if <division> is 96, then a time
-    /// interval of an eighth-note between two events in the file would be 48.
-    QuarterNote(u16), // TODO - clamp this to the allowable range
-    /// Frame rate and resolution within the frame.
-    Smpte(SmpteRate),
-}
-
-impl Default for Division {
-    fn default() -> Self {
-        Division::QuarterNote(1024)
-    }
-}
-
-const DIVISION_TYPE_BIT: u16 = 0b1000000000000000;
-
-impl Division {
-    pub(crate) fn from_u16(value: u16) -> LibResult<Self> {
-        if value & DIVISION_TYPE_BIT == DIVISION_TYPE_BIT {
-            // TODO - implement SMPTE division
-            error::Other { site: site!() }.fail()
-        } else {
-            Ok(Division::QuarterNote(value))
-        }
-    }
-
-    pub(crate) fn write<W: Write>(&self, w: &mut W) -> LibResult<()> {
-        match self {
-            Division::QuarterNote(q) => Ok(w.write_all(&q.to_be_bytes()).context(wr!())?),
-            Division::Smpte(_) => error::Other { site: site!() }.fail(),
-        }
-    }
-}
-
-impl TryFrom<u16> for Division {
-    type Error = Error;
-
-    fn try_from(value: u16) -> Result<Self> {
-        Ok(Division::from_u16(value)?)
-    }
-}
-
-/// <division> Bits 14 thru 8 contain one of the four values -24, -25, -29, or -30, corresponding to
-/// the four standard SMPTE and MIDI time code formats (-29 corresponds to 30 drop frame), and
-/// represents the number of frames per second. These negative numbers are stored in two's
-/// complement form.
-#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd, Hash)]
-pub enum FrameRate {
-    /// 24 frames per second
-    N24,
-    /// 25 frames per second
-    N25,
-    /// 30 drop
-    N29,
-    /// 30 frames per second
-    N30,
-}
-
-impl Default for FrameRate {
-    fn default() -> Self {
-        FrameRate::N24
-    }
-}
-
-#[derive(Clone, Copy, Debug, Eq, Ord, PartialEq, PartialOrd, Hash)]
-pub struct SmpteRate {
-    /// The number of frames per second.
-    frame_rate: FrameRate,
-    /// The <division> second byte (stored positive) is the resolution within a frame: typical
-    /// values may be 4 (MIDI time code resolution), 8, 10, 80 (bit resolution), or 100. This system
-    /// allows exact specification of time-code-based tracks, but also allows millisecond-based
-    /// tracks by specifying 25 frames/sec and a resolution of 40 units per frame. If the events in
-    /// a file are stored with bit resolution of thirty-frame time code, the division word would be
-    /// E250 hex.
-    resolution: u8,
-}
-
-impl Default for SmpteRate {
-    fn default() -> Self {
-        // This is the 'millisecond-based tracks' example given by the spec.
-        SmpteRate {
-            frame_rate: FrameRate::N25,
-            resolution: 40,
-        }
-    }
 }
 
 #[derive(Clone, Debug, Default, Eq, Ord, PartialEq, PartialOrd, Hash)]
